+++ frontend/src/AnnouncementsDisplay.css
@@ -1,0 +1,12 @@
+.announcements-display {
+    position: fixed;
+    bottom: 50px;
+    right: 5px;
+    width: 500px;
+    border: 1px solid #ddd;
+    margin: 5px;
+    padding: 5px;
+    border-radius: 5px;
+    background-color: #cccccc;
+    font-size: 1em; /* Smaller font size */
+}
